Working analysis options editor
