+++ src/app/angular/components/analysis/advanced/advanced.component.html
@@ -1,0 +1,58 @@
+<div class="container">
+  <div class="pt-3 pb-3">
+    <div class="card shadow">
+      <div class="card-header">
+        Advanced Options
+      </div>
+      <div class="card-body d-flex justify-content-center">
+        <button class="btn btn-warning">
+          Enable Advanced Features
+        </button>
+      </div>
+
+      <ng-content></ng-content>
+    </div>
+  </div>
+  <div class="pt-3 pb-3">
+    <div class="card shadow">
+      <div class="card-header card-header-form">
+        <div class="header">Analysis Options</div>
+        <div class="flex-fill"></div>
+        <form class="form-inline my-0">
+          <button
+            class="btn btn-outline-danger my-0 py-1"
+            (click)="resetOptions()"
+          >
+            Reset
+          </button>
+        </form>
+      </div>
+      <div class="card-body">
+        <app-option
+          *ngFor="let option of options"
+          [analysis]="analysis"
+          [option]="option"
+        ></app-option>
+      </div>
+    </div>
+  </div>
+  <div class="pt-3 pb-3">
+    <div class="card shadow">
+      <div class="card-header card-header-form">
+        <div class="header">Configuration Editor</div>
+        <div class="flex-fill"></div>
+        <form class="form-inline my-0">
+          <button
+            class="btn btn-outline-danger my-0 py-1"
+            (click)="resetConfig()"
+          >
+            Reset
+          </button>
+        </form>
+      </div>
+      <div class="card-body">
+        <app-config-option [config]="config"></app-config-option>
+      </div>
+    </div>
+  </div>
+</div>
