+++ src/app/angular/components/analysis/audio/audio.component.html
@@ -1,6 +1,6 @@
 <div class="container mt-3">
   <div class="card shadow">
-    <div class="card-header d-flex">
+    <div class="card-header card-header-form">
       <div class="header">Audio Files</div>
       <div class="flex-fill"></div>
       <form class="form-inline my-0">
