+++ src/app/angular/components/analysis/audio/audio.component.scss
@@ -1,21 +1,3 @@
-.card {
-  .card-header {
-    padding-bottom: 0;
-    padding-right: 0;
-    padding-top: 0;
-
-    .btn {
-      margin-right: 0.25rem;
-      margin-top: -1rem;
-    }
-
-    .header {
-      padding-bottom: 0.5rem;
-      padding-top: 0.5rem;
-    }
-  }
-}
-
 table {
   width: 100%;
 }
