--- src/app/angular/components/analysis/folder/folder.component.scss
@@ -1,17 +1,0 @@
-.card {
-  .card-header {
-    padding-bottom: 0;
-    padding-right: 0;
-    padding-top: 0;
-
-    .btn {
-      margin-right: 0.25rem;
-      margin-top: -1rem;
-    }
-
-    .header {
-      padding-bottom: 0.5rem;
-      padding-top: 0.5rem;
-    }
-  }
-}
