+++ src/styles.scss
@@ -11,3 +11,20 @@ body {
 
   height: 100%;
 }
+
+.card-header-form {
+  display: flex;
+  padding-bottom: 0;
+  padding-right: 0;
+  padding-top: 0;
+
+  .btn {
+    margin-right: 0.25rem;
+    margin-top: -1rem;
+  }
+
+  .header {
+    padding-bottom: 0.5rem;
+    padding-top: 0.5rem;
+  }
+}
